Moved custom slides to own app topics for better app structure. Renamed model to Topic. Added migrations file. Fixed #2402.
