+++ CHANGELOG
@@ -18,6 +18,7 @@ Assignments:
 Core:
 - Added support for big assemblies with lots of users.
 - Added HTML support for messages on the projector.
+- Moved custom slides to own app "topics". Renamed it to "Topic".
 
 Motions:
 - Added origin field.
