--- openslides/core/static/templates/core/slide_customslide.html
@@ -1,4 +1,0 @@
-<div ng-controller="SlideCustomSlideCtrl" class="content scrollcontent">
-  <h1>{{ customslide.agenda_item.getTitle() }}</h1>
-  <div ng-bind-html="customslide.text | trusted"></div>
-</div>
