+++ openslides/topics/static/templates/topics/slide_topic.html
@@ -1,0 +1,4 @@
+<div ng-controller="SlideTopicCtrl" class="content scrollcontent">
+  <h1>{{ topic.agenda_item.getTitle() }}</h1>
+  <div ng-bind-html="topic.text | trusted"></div>
+</div>
